Add files via upload
My first photo for this activity depicts a large jar with lots of paintbrushes and different tools an artist would use while working. I felt this photo related to my current career status in the way that I am the Jar and am constantly adding knew tools, experience, and knowledge of them to my collection. My second photo depicts a wealthy individual with money falling from his pockets and other hands reaching for it, symbolizing to me the success and wealth I am going to acquire with my tools. The way that I am going to bridge the gap between these two photos is by actively adding to my collection of skills and tools while learning and applying them in a business environment and mindset.